Initial Commit
4/11/23 - 6:14 PM EST
Right before class, initial data functions with routes and no actual values yet.

Need to know if I need to set my stuff to a certain route, because you definitely need to update the library with that route.

I use prettier, so you'll notice a lot of code cleanup. Check my comments in data for suggestions about the broader code.